tests: Access drives via fake vm
Pass the fake vm to all the helper function that need to access the
drive list. This will make it possible to create different drive
configuration, for example block based drives that need to be extended
during backup.

Now that the fake vm is available in the helper function, we can use its
id attribute instead of duplicating it.

Change-Id: I5c67319b0f27a0ae74c586e3b90dff565e843d85
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>