tests: Generate backup xml automatically
When we start a backup, we need to parse libvirt backup xml to get the
index for the backup scratch disk. This requires that we have a backup
xml for every test running backup.start_backup(). We have 22
invocations, so setting this manually is not the way.

FakeDomainAdapter generates now the backup xml from the backup_xml
argument to backupBegin(). Since we always have backup xml, we can
verify that backup was started correctly by comparing the backup xml
instead of keeping and comparing the input xml.

Similar change is needed for verifying checkpoint xml.

Change-Id: I64a5202b8f1ef3cad0092c6949c0a32c65fb9f10
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>