backup: Use exportname to create backup url
Libvirt exposes now the exportname via the backup xml[1]. We can set
custom export name when starting a backup, but we don't do this yet, so
we always get the drive name. Using the exportname make the code future
proof in case we want to use a custom export name.

Here an example parsed backup xml from the tests:

    {
      'incremental': '90f469ad-1ed2-46fc-9851-0d5db9b8515c',
      'socket': '/run/vdsm/backup/db250ab6-ce07-4d83-ae80-6b8c5557e956',
      'disks': {
        'sda': {
          'index': 9,
          'exportname': 'sda',
        },
        'vda': {
          'index': 10,
          'exportname': 'vda',
        },
      },
    }

[1] https://libvirt.org/formatbackup.html

Change-Id: I7d9303cf121be79a58a048ca2cc8175481e67106
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>