backup: Parse disk type from backup xml
Based on the disk type, we can decide if scratch disk needs to monitored
and extend. When we start backup we already have this info, but getting
the value from libvirt means we can start monitoring disks anytime, for
example after vdsm is restarted, based on libvirt state.

Here is an example parsed backup xml:

    {
      'incremental': '7554d598-edaf-49ba-a476-fde54ca3c7fd',
      'socket': '/run/vdsm/backup/e8ef325d-058e-416c-b9a2-8b9b90da811b',
      'disks': {
        'sda': {
          'index': 9,
          'exportname': 'sda',
          'type': 'file',
        },
        'vda': {
          'index': 10,
          'exportname': 'vda',
          'type': 'file',
        },
      },
    }

Change-Id: I27869805102c6df501da64272659bec3dc57bbda
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>