backup: Improve naming
The backup module is reusing the term "drive" for backup info for
particular drive. This confuses the reader, expecting that "drive" is a
virt.vmdevices.storage.Drive object.

Rename BackupDrive to BackupDisk, and use "backup_disk" in places when
we used "drive". Now "drive" is used only when accessing a real Drive
object.

Change-Id: I56277a6e91fd4d7dc77436b4a53e9014f76c732a
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>