backup: Keep scratch disk info in drive
When using block based scratch disk, keep the scratch disk info in the
Drive object during the backup.

Currently we have only the index, which can be used to register block
threshold events. We need to modify engine to send also the scratch disk
domain, image, and volume ids so we can also extend them.

Change-Id: I7df8312386da2bc628efc7bf2fba6669c59f81d0
Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>