lvm: Don't use global:locking_type
This option is deprecated and useless since RHEL 8. We could not remove
it in the past since we had to support older version of lvm on Fedora.

Looks like this option became harmful in lvm2-2.03.14-1.el8.x86_64,
converting locking_type=4 to --readonly. We see this failure in vdsm
tests:

    WARNING: locking_type (4) is deprecated, using --sysinit --readonly.
    Operation prohibited while --readonly is set.
    Can\'t get lock for b4512b9d-84dc-43ba-865d-32c4a1cd148a.

and the lvm command fails.

Converting locking_type=4 to --readonly does not look correct, so this
is likely a regression in lvm. But we should not use locking_type in
vdsm. This option is used only in tests using read only mode, which is
never used in the real application.

As a quick fix to unbreak the tests, remove the locking_type
configuration. We need to remove the tests for read only mode and remove
the entire read only mode feature later.

Change-Id: Ia9af81756c07c26517805633af5d90d523e60fe7
Bug-Url: https://bugzilla.redhat.com/2025527
Signed-off-by: Nir Soffer <nsoffer@redhat.com>