virt: Prevent migrations of VMs paused due to an I/O error
Migrations of such VMs will be rejected by libvirt, it doesn't make
sense to start the migration machinery for them.  It's also good to
avoid resuming such VMs during migration preparation if possible,
which may fail and prevent the VM from resuming later or which may
reveal unhandled races or corner cases.

Migration of such VMs is already prevented in Engine but let's add an
additional check to Vdsm to handle races, similarly to _not_migrating
API guard.

This patch doesn't handle the case when the VM gets paused and
resumed when the migration process is already running. It will be
addressed in another patch.

Change-Id: Iec4e343f6c3cf39f36b339987d27c9c32b40c0a4
Bug-Url: https://bugzilla.redhat.com/2010478
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>