virt: Don't assume that VM is UP after a failed migration
Current migration code switches VM status to UP
unconditionally (unless in recovery) after a failed migration.  But
this is wrong, the migrated VM could be paused initially, it could be
paused due to a storage error, it could crash, etc.

Tracking these somewhat corner case during migration would be
difficult.  The migration either succeeds and then the VM is simply
down or it fails.  In the latter case, it's easiest to use our status
recovery routine and ask libvirt about the true status of the VM.
We additionally ensure that we won't get stuck in the
MIGRATION_SOURCE status by still switching to UP in such a case.

Change-Id: I30fbbf71dea9c32cb0af8910111c7483d2475144
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>