virt: Postpone VM resume if blocked due to the VM status
When a storage domain status changes, it's attempted to resume all the
VMs currently paused because of I/O error.  If a given VM is down,
hibernating or migrating then the resume action is not performed.
This is not a problem when the VM is already down or when it is
hibernating (it will either get down on success or get resumed on
failure).  But if the VM is migrating to another host, the migration
will fail (libvirt cannot migrate VMs paused due to I/O error)
then the chance to resume the VM gets missed and the VM may remain
paused although it should be resumed.

Let's fix the problem by storing a flag about a refused resume attempt
and resuming the VM after a failed migration.

Under some circumstances, e.g. when a VM is waiting on
ongoingMigrations semaphore, it may be better and possible to resume
the VM immediately.  This will be addressed in a separate patch.

Change-Id: Ie6458c9c2567c432fb13071d6cb7fb0bbd8de732
Bug-Url: https://bugzilla.redhat.com/2010478
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>