copy_data: skip locking source when same image is used
When we copy a volume and target image is the same as the source volume we must
skip locking the source with a shared lock and only lock the target with
an exclusive lock.
Locking both would lead to a deadlock and fail the operation.

This patch checks if both source and target are the same image and skips or
locks accordingly.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>
Bug-Url: https://bugzilla.redhat.com/977778
Change-Id: I4b71830b86b961acc73fd26975c14af2102fe270