storage: use specific class for connection
Instead of using parent class for connecting prepared connections,
return child class from prepare connection function and use it for
connecting. This finally delegates connection to sub-class and after
this change we can overide connect function in sub-class and implement
parallel connection for given sub-class.

Mixing different types of connection is not supported, so we don't have
to to care about this case.

Change-Id: Iee4d531496a3b1754db42941d8e0c9364d64dbdf
Bug-Url: https://bugzilla.redhat.com/1787192
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>