ci: Enable timestamp in rpm names
For merged patches we want the standard git describe string as used in
vdsm copr builds[1].

For pull requests or local builds enable the timestamp that makes it
easy to upgrade a local development build.

The timestamped builds also solve the issue when OST fail to run your
patch because it is considered older than the latest build. With this
change, a patch built in github is always newer than the latest build
with the same version (e.g. v4.50.0.2).

Example merged patch build:
https://github.com/nirs/vdsm/runs/4408896025

Example pull request build:
https://github.com/nirs/vdsm/actions/runs/1535544653

[1] https://copr.fedorainfracloud.org/coprs/ovirt/ovirt-master-snapshot/package/vdsm/

Change-Id: Ib4160b4e9697a38205fd6f02aed24c8604b29c68
Signed-off-by: Nir Soffer <nsoffer@redhat.com>