userstorage: Retry creating loop device
Creating a loop device with the --sector-size option may fail randomly
if the device has dirty pages from previous usage. This was fixed in
losetup from util-linux 2.37.1, but this version is not available in
Centos Stream 8.

Fixed by adding a retry loop, similar to the retry loop used internally
in losetup.

Here is example random failure, fixed on the first retry:

    [userstorage] WARNING Attempt 1/20 failed: losetup: /dev/loop5: set
    logical block size failed: Resource temporarily unavailable

Change-Id: I285dedd09abd89e62152b887a3b05807c627041e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>