common: Make sure passwords remain protected
unprotect_passwords modifies the given object in place.  This is
dangerous and may reveal passwords in log.  Consider the following
scenario:

  log.info("Data: %s", object_with_passwords)
  revealed = unprotect_passwords(object_with_passwords)

Logging operations may be asynchronous and at the time the parameter
is written to the log, it may be already modified and contain
unprotected passwords.

We already return the modified object from unprotect_passwords so
there is no real need to modify the original object.  Let's modify its
copy instead.

Similar issues may be with protect_passwords but there we are on the
safe side as for password security.

Change-Id: If9ba163dc05a081ebe124900d61c038e2c21fea2
Bug-Url: https://bugzilla.redhat.com/2033697
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>