virt: Permit resuming migrating VMs stopped due to I/O errors
We block resuming VMs in MIGRATION_SOURCE status.  For good reasons,
VMs are paused in the final stages of migrations and we don't want to
mess with this and generally VMs states during migrations.

But it also means that if a VM gets temporarily stopped due to an I/O
error while waiting for migration, it cannot be resumed if the storage
problem disappears while still waiting for the migration.  If the VM
is waiting on ongoingMigrations semaphore then it may keep it paused
unnecessarily for long time.

We can improve this situation by checking whether the VM in
MIGRATION_SOURCE status is paused on I/O error and resuming it in such
a case.  libvirt prevents migrating such VMs and we can be sure that
we don't interfere with a running migration when resuming such a VM.
We make the check by asking libvirt directly, to be completely sure
about the VM status.

Change-Id: I432f51b1d7490d1cdc0d8ea484a96eb74d5da084
Bug-Url: https://bugzilla.redhat.com/2010478
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>