storage: move udev settle into dedicated method
After connecting to iSCSI server we wait for udev to process all new
devices up to specified timeout (by default 5 seconds). This is called
also from IscsiConnection._maybe_connect_iser() for no good reason as
we only connect and disconnect over iSER. udev settle needs to be called
only when we finish with connecting to the iSCSI server. Move it into
dedicated function and call it as a last step.

Change-Id: Ifbcac50a4560b1e59ab21019c33709aa89a919c6
Bug-Url: https://bugzilla.redhat.com/1787192
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>