automation: Remove all tests
Remove all the tests that fail randomly in gerrit - we have fast and
reliable CI in github now. The tests jobs only overload the CI and cause
other jobs to fail.

Remove all the build artifacts jobs except x86_64 el8stream/rhel8, used
by OST. Currently build artifacts jobs fail randomly in one of the
nodes.

The actual test files in automation/ were not removed yet, to make it
easy to enable the tests if we have a need. We will remove the files
once we move to github completely.

Change-Id: I0cb65f43645efb37e83c30931091d93e330da747
Signed-off-by: Nir Soffer <nsoffer@redhat.com>