storageServer: Improve logging for iscsi connection
We typically log info message for every operation that modifying the
system state. Discovering iscsi targets, adding and removing iscsi nodes
from iscsi db and logging in and out from iscsi targets are such
operations but had no logs.

The new logs should make it easier to debug and support the system, and
will likely reveal bugs we are not aware of. The new logs show only
when connecting and disconnecting from iscsi targets and when
discovering new targets.

Change-Id: I5c3fd97a2c4033de4a52a6296c3e1ea36753fb12
Signed-off-by: Nir Soffer <nsoffer@redhat.com>