Simplify VM._amend_block_info()
This function is needed to amend block info when replicating a
non-chunked drive to chunked drive, but it was abused to set
Drive.blockinfo, used to log changes in extension info.

Convert Drive.blockinfo to a property, and move the logging code to the
property setter. Set Drive.block_info after separately from the call to
amend the block info.

This change allows amending block info only for drives during
replication, which will make it easier handle scratch disks in the
monitoring flows.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>