tests: Adapt nbd tests to qemu-nbd 6.2
Looks like zero detection when writing to qcow2 image has changed, and
now the zeroes are converted to zero clusters. This looks like a nice
bug fix compared with the previous behavior.

Update the test so it does not check actual image size for qcow2
format, which is to fragile for testing anyway.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>