nbd: Disable allocation-depth for raw format
qemu-nbd 6.2.0 introduced a bug[1] when accessing raw image extents. The
first call works correctly, but the next calls the entire image is
reported as data.

I could not reproduce the issue using imageio API, so we are likely not
affected by this bug. However this reveal that we were not careful
enough in qemu-nbd configuration. We always enabled --allocation-depth
when it is useless for raw format, always reporting depth=1.

Change to enable --allocation-depth only for qcow2 format when it
has meaningful output.

[1] https://bugzilla.redhat.com/2041480

Signed-off-by: Nir Soffer <nsoffer@redhat.com>