lvm: refactor _buildFilter() function
To be able to replace filter with devices in lvm commands, start with
refactoring _buildFilter() function. Factor out part which construct
list of allowed devices. This can be reused also for construcing
`--devices` argument for lvm commands.

This is a preparation step for caching devices instead of filter.
Switch to caching devices will be done in follow-up patch.

Change-Id: I25544f92cbcfeadace6a501412f94f43a98ef392
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>