lvm: cache devices instead of filter
To be able to use `--devices` argument in lvm commands, we need only
list of devices, not filter. On the other hand, filter can be easily
constructed from device list. Cache only allowed devices and use it
for constructing lvm filter. Also rename related methods to reflect
we cache devices and not filter.

Change-Id: I59a418a66d418305c9f9afd0f2b3ef87b421861e
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>