github: Fix can_push condition
When we set:

    can_push: ${{ github.repository_owner == 'oVirt' }}

Github converts the boolean value to the strings 'true' or 'false'[1],
so checking:

    if: ${{ env.can_push == true }}

Does not match and now we never push to quay.

Example run on my fork show that we still skip the push on forks:
https://github.com/nirs/vdsm/actions/runs/1740365670

Unfortunately, the only way to test this properly is to merge to master.

[1] https://docs.github.com/en/enterprise-server@3.0/actions/learn-github-actions/expressions#functions

Fixes 746975bb434e5fbb4c41c3e119728bc3322fb5e9

Signed-off-by: Nir Soffer <nsoffer@redhat.com>