virt: Rename drive monitoring to volume monitoring
In the past we could monitor only the active volume of a drive, so drive
monitor was the right term.

However, now we can monitor also:
- Scratch disk volume
- Internal volume during live merge (blockCommit)
- Target volume during disk replication (blockCopy)

Modifying the Drive object to keep monitoring info is hard and does not
look like the right way. The plan is to keep monitoring info in the
volume monitor.

When we want to monitor a volume, we will call:

   VolumeMonitor.start_monitoring(volume)

This will set a block threshold on the volume using libvirt, and keep
the volume state needed to handle block threshold events.

When we want to stop monitoring a volume, we will call:

   VolumeMonitor.stop_monitoring(volume)

This will disable block threshold events and remove the volume from
the monitor.

When we get a block threshold event, the monitor will update the volume
state.

In the periodic job, we will query the volumes that need to be extended
and extend them.

When a drive changes the path, for example at the end of live storage
migration or live merge, we will remove the old volume from the volume
monitor, and add the new volume.

Starting this work with renaming to make the intent of the code clear:

- drivemonitor module to thinp - this module will include all the code
  needed for handling thin provisioned block volumes.
- drivemonitor.DriveMonitor class to thinp.VolumeMonitor.
- VolumeMonitor.monitored_drives to VolumeMonitor.monitored_volumes.
- VM.drive_montior to VM.volume_monitor.
- VM.montior_drives to VM.montior_volumes
- periodic.DriveWatermarkMonitor to periodic.VolumeWatermarkMonitor.
- "drive monitor" to "volume montior" in doc/thin-provisioning.md.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>