vm: Rename _resume_if_needed and make it public
We have very confusing _resume_if_needed() and maybe_resume() - which
method should be called to resume a VM? Rename _resume_if_needed() to
extend_volume_completed(), live merge _extend_completed callback. Now
it is clear that this should be called only when extend flow completed,
and other code should call maybe_resume().

This new method is public so we can move the extend code to the thinp
module.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>