lvm: disable devices file when searching for mounts
Disable devices file when searching for mounts and use filter. This is
correct as either called when vdsm is configured to use filter and
threfore devices file should be diaabled or vdsm is configured to use
devices file, but it's not configured yet.

This allows us to avoid lvm warnings that we mix devices file and lvm
filter. Such situation happends when use configured host with devices
files and later wants revert configuration back to filter.

Change-Id: I634c6470f8ccfa1300c105ba2bf4f2e03316e16f
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>