vm: Merge methods for refreshing destination
We had 2 confusing methods for refreshing the destination volume during
migration. Since both of them are small and simple, and the private
method should not be called in any other flow, merging them makes the
code easier to follow.

While merging the methods, streamline the way we raise on errors.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>