lvm: always disable devices file when filter should be used
If there are any issues with devices file on the host or user wants to
use lvm filter for any other reason, always disbale devices file in lvm
config. It can be turned on from previous run of vdsm config or it needs
to be overridden on CentOS 9 where devices file is enabled by default.

Change-Id: I14d2f606f4f2dbd7ce485fdbc9b3270ce09a72d8
Signed-off-by: Vojtěch Juránek <vjuranek@redhat.com>