vm: Move extending code to thinp
Move the complex extending code to the thinp module. This will make it
easier to monitor volumes instead of drives. The VM is not affected now
by the way we monitor or extend volumes.

The VM part during disk extension is now:

- Call VolumeMonitor.monitor_volumes() periodically. This is done by
  the periodic module.

- Call VolumeMonitor.extend_volume() when volumes need to be
  pre-extended. This is called when starting disk replication, and
  before starting the commit phase in live merge. This will be
  eliminated when we start to monitor the relevant volume.

- Handle callbacks from VolumeMonitor._after_volume_extension:
  - Refresh migration volume
  - Refresh local volume
  - Resume if needed when extend volume completed

Because extending volumes is implemented now by the volume monitor, we
have to use the real object in live merge test.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>