thinp: Unify callback names
Unify the callbacks name with other callbacks. The callback is named
{method_name}_completed to make it easy to find the callback when you
look at the method, or the other way around.

The storage system refereed to the private callback for explaining why
we don't refresh extended volumes on the SPM. This is bad, requiring
unrelated changes when renaming private methods. Change the text to
explain what we do without referring to the actual method name.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>