vm: Move amend_block_info() to thinp
This method is part of extending flow. I kept it in the VM because it
uses VM.getVolumeSize(), but the volume montior already depends on it,
and is not really related to the VM.

This is another ugly hack that will be removed when volume monitoring is
used for all volumes. Until we remove it, it should live in thinp.

Bug-Url: https://bugzilla.redhat.com/1913387
Signed-off-by: Nir Soffer <nsoffer@redhat.com>