thinp: Separate block info update from extend check
On each monitor cycle, we need to get block stats from libvirt and
updated the block info for every drive that needs monitoring. Based on
the block info, we decide if the drive should be extended.

Separate the block info update from the extend check. This makes the
code easier to change and understand.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>