tests: Fix ancient invalid test
Remove evil faking of VolumeMonitor._update_threshold_state_exceeded().
The evil faking was added when we fixed the monitor to set a drive as
exceeded. This broke existing test, so the author faked the call for the
test to make the test pass. I guess this was done as a temporary work in
progress hack, but somehow it got merged.

To fix the test, we added support for specifying replica size. The size
is the size of the logical volume used as replica. This enables
simulating the case when the original volume is on file storage, but the
replica volume is on block storage.

Since the tests setup is correct, we don't try to extend the volume, and
the test pass without modifying the private method.

Fixes: c284b8202f0d151ef6b7a592baaa4d8490a14607
Signed-off-by: Nir Soffer <nsoffer@redhat.com>