tests: Test monitoring_needed()
This is a very simple method, but it was not tested before. In the past
we tested _monitored_volumes(), which was very complicated, but after
all the logic moved to Drive.needs_monitoring(), it became trivial.

All the interesting and important cases we tests are already in Drive
tests, so there is no need to test them again here. Simplify the test by
creating a trivial FakeDrive class implementing only needs_monitoring().

Testing monitoring_needed() tests also _montior_volumes(), enable() and
disable().

This change deletes 300 lines, but test coverage is still 76%.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>