common: Support measuring since time in the past
There are cases when want to measure time since time in the past. For
example, when we measure the time to extend a volume, we want to measure
the total time since the time we received a block threshold event.

Add optional start_time to Clock.start(). When specified, the clock will
measure that time since the specified start time, instead of the time
the clock was started.

A new test show an example for this use case.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>