thinp: Include the wait time in total extend time
If we received a block threshold event, add new "wait" step, measuring
the time since the block threshold event was received until we handled
it.

The "total" time shows now the time since we found that the drive was
exceeded, until the new size is visible to the guest. This time is
important for setting up good chunk size and utilization options.

Example extend timing with this change:

2022-02-21 12:32:45,915+0200 INFO  (mailbox-hsm/1) [virt.vm]
(vmId='d0730833-98aa-4138-94d1-8497763807c7') Extend volume
3b435e98-f6b1-40ab-b3de-18b403049e7a completed <Clock(total=3.85,
wait=1.55, extend-volume=2.10, refresh-volume=0.20)> (thinp:567)

Signed-off-by: Nir Soffer <nsoffer@redhat.com>