tests: Test drive exceeded time behavior
The way we set and clear exceeded time is little too delicate. Lets
protect with it with a test.

The new test added as simple function since the relevant class was was
not converted properly to pytest, and we cannot use pytest fixtures in
the class. Converting the module to pytest properly will have to be done
later.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>