tests: Fix merge test mocking
This horrible test was depending on default configuration instead of
using the config, and is written in a way that make it hard to use the
config.

The horrible make_env() context manager using the crappy storagetestlib
was mocking everything after creating the volumes that need mocking to
consider the configuration. Change to make create everything inside the
mock context.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>