thinp: Simplify monitor_volumes()
Previously we returned True if some drives were extended, and ENOSPC
handler logged a message if some drives were extended. This log is not
very useful.

Simplify the interface by not returning anything. This will allow
handling the monitoring request in the executor thread pool, and it
simplifies the monitoring code.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>