volume: Support adding a bitmap to new volume
When creating a new volume or a snapshot, caller can specify
bitmap={bitmap-uuid} to create a new empty bitmap in the new volume.

This will enable hybrid backup, when every backup is started by creating
a temporary snapshot. The new bitmap records the changes to the volume
since the snapshot was created. When the temporary snapshot will be
deleted, the new bitmap will be copied to the parent volume, and will be
used for the next incremental backup.

Like add_bitmaps, we treat creating a new bitmap as best effort. If the
operation fails we don't fail the creation of the volume, but the next
backup using this bitmap will have to be a full backup.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>