virt: Add support for vGPU driver parameters
Engine is going to add a new metadata item for vGPU devices,
mdevDriverParameters.  It specifies driver parameters to be set when
the mediated device is created, by writing a file in /sys.

Since mediated devices are created before each use and destroyed after
each use, it's enough to write the driver parameters, if specified,
only when the device is created.  It's not necessary to clear them
when the device is removed, although it is possible to do it by
writing a space to the driver parameters file in /sys.

Currently, only NVidia driver parameters are supported.  If setting
driver parameters is attempted for other cards or if setting the
driver parameters fails for another reason, creation of the mediated
device will fail and the VM won't be started.

Bug-Url: https://bugzilla.redhat.com/1987121