nbd: Always verify bitmap
Previously we checked if the bitmap exists only if the volume had a
parent, when we create an overlay. If we export a single volume and the
bitmap was missing, qemu-nbd starts normally, but exit immediately when
it try to export the bitmap. Since we detected that qemu-nbd started,
NBD.start_server() succeeds, and engine report the transfer as ready.
When the user try to connect to imageio server, it fails since qemu-nbd
is not running.

Here is an example failure:

    Mar 08 01:44:43 host4 systemd[1]: Started /usr/bin/qemu-nbd --socket
    /run/vdsm/nbd/54653918-63b2-49c1-8e7f-3c2d47db2bfc.sock --persistent --shared=8
    --export-name= --cache=none --aio=native --allocation-depth --read-only
    --bitmap=859e94f0-aaab-4a3b-bd65-a7dc1b8d822b json:{"driver": "qcow2", "file":
    {"driver": "host_device", "filename": "/rhev/...-4c9808da12e8"}}.

    Mar 08 01:44:43 host4 qemu-nbd[27200]: qemu-nbd: Bitmap
    '859e94f0-aaab-4a3b-bd65-a7dc1b8d822b' is not found

Fix the issue by verifying that the bitmap exists in all cases, and
failing with the same error we fail when a bitmap is missing or invalid
in the backing chain. This will fail the transfer earlier, which will
make it easier to debug the issue.

Regardless of this fix, we need to understand why qemu-nbd starts
normally when the bitmap is missing. This smells like a bug in qemu-nbd,
creating a server socket before it validates the bitmap argument.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>