nbd: Create overlay only if needed
If we export a volume with a parent, but the bitamp does not exist in
the parent, we don't need to create an overlay. The overlay is needed
only when the bitmap exists in the volume backing chain, since qemu-nbd
export the bitmap only from the top volume.

Now we create an overlay only if the bitmap chain has more than one
item. If it contains only one item this must be the exported volume
(enforced by _find_bitmap).

Signed-off-by: Nir Soffer <nsoffer@redhat.com>