tests: Tests backing chain with file
Measure without a backing chain was tested only with block storage.
Add tests for file storage and refactor the code creating the chain from
the actual test.

Due to the way we create block storage, we cannot use yet the same
storage with multiple tests, but if we change the way storage is defined
we can use a module or session scope fixtures.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>