qemuimg: Support measuring sub chain
Add new "base" argument, enabling measuring of a sub chain, without the
rest of the backing chain:

    parent <- [base <- top] <- child

The typical use case is estimating the size of the base volume after
merging the top volume into the base.

This will allow extending the base volume to the right value before
merge instead of over allocating, that cannot be fixed in live merge.

The API will allow engine to measure volumes before merge instead of
guessing in a bad way, assuming vdsm internal details like chunk size
and qcow2 metadata overhead.

Currently we support only chains with 2 elements, since we don't support
merging of longer chains. If support for longer chains will be added in
the future, we can remove the limit in the measure implementation.

Vdsm reports new "measure_subchain" capability to allow engine to
measure sub-chain with older cluster versions.

Bug-Url: https://bugzilla.redhat.com/2064907
Signed-off-by: Nir Soffer <nsoffer@redhat.com>