virt: allow modifying manual CPU pinning during migration
The original implementation was pre-creating XML elements early on the
VM.migrate call. The rationale behind it being to validate the passed
parameter and avoid as many errors as possible later in the migration
thread. This however makes it useful only for CPU policies that do not
have manual pinning (otherwise it would lead to duplicate elements in
final XML).

To make it possible to reconfigure also manual CPU pinning during
migration the code was split. The parameter verification is still done
early but the XML is constructed later in the source thread as needed.

Change-Id: Ic4b50d59ea7bbb8e9b814852044934cbca048550
Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>