mailbox: Minimize messages latency
Minimize message latency by introducing mailbox events mechanism. The
unused host 0 mailbox is used now for sending and receiving mailbox
events.

To control mailbox events, a new "mailbox:events_enable" was added. The
option is disabled by default, so we can test this change before we
enable it by default, or disable it in production if needed.

To enable mailbox events, add a drop-in configuration file to all hosts:

    $ cat /etc/vdsm/vdsm.conf.d/99-local.conf
    [mailbox]
    events_enable = true

And restart the vdsmd service.

When mailbox:events_enable option is enabled:

- Hosts write an event to host 0 mailbox after sending mail to the SPM.

- The SPM monitors host 0 mailbox every eventInterval (0.5 seconds)
  between monitor intervals, so it can handle new messages quickly.

- When hosts wait for reply from the SPM, they monitor their inbox every
  eventInterval (0.5 seconds), so they detect the reply quickly.

- Host reports a new "mailbox_events" capability. This can be used by
  engine to optimize mailbox I/O when all hosts in a data center
  supports this capability.

With this change, extend roundtrip latency reduced from 2.0-4.0 seconds
to 0.5-1.0 seconds, reducing the risk of pausing VM when writing quickly
to fast storage.

Testing shows that we can write now 525 MiB/s (the maximum rate on my
nested test environment) in the guest without pausing a VM when the disk
is extended. Before this change, we could write only 350 MiB/s before
the VM starts to pause randomly during the test.

Here are example logs form this run, showing that total extend time is
1.14-3.31 seconds instead of 2.5-8.3 seconds before this change.

    <Clock(total=1.65, wait=0.28, extend-volume=1.09, refresh-volume=0.28)>
    <Clock(total=2.67, wait=1.80, extend-volume=0.58, refresh-volume=0.29)>
    <Clock(total=3.10, wait=1.74, extend-volume=1.10, refresh-volume=0.25)>
    <Clock(total=2.85, wait=1.55, extend-volume=1.08, refresh-volume=0.22)>
    <Clock(total=2.02, wait=1.14, extend-volume=0.58, refresh-volume=0.30)>
    <Clock(total=1.14, wait=0.33, extend-volume=0.56, refresh-volume=0.25)>
    <Clock(total=2.83, wait=1.42, extend-volume=1.09, refresh-volume=0.32)>
    <Clock(total=1.68, wait=0.33, extend-volume=1.10, refresh-volume=0.25)>
    <Clock(total=2.45, wait=1.47, extend-volume=0.60, refresh-volume=0.38)>
    <Clock(total=1.44, wait=0.11, extend-volume=1.09, refresh-volume=0.24)>
    <Clock(total=2.46, wait=1.04, extend-volume=1.13, refresh-volume=0.30)>
    <Clock(total=1.55, wait=0.17, extend-volume=1.07, refresh-volume=0.31)>
    <Clock(total=2.02, wait=1.13, extend-volume=0.60, refresh-volume=0.28)>
    <Clock(total=1.75, wait=0.39, extend-volume=1.12, refresh-volume=0.24)>
    <Clock(total=2.98, wait=1.61, extend-volume=1.12, refresh-volume=0.25)>
    <Clock(total=1.28, wait=0.41, extend-volume=0.61, refresh-volume=0.25)>
    <Clock(total=2.46, wait=1.48, extend-volume=0.62, refresh-volume=0.36)>
    <Clock(total=3.31, wait=1.97, extend-volume=1.12, refresh-volume=0.21)>
    <Clock(total=1.81, wait=0.96, extend-volume=0.58, refresh-volume=0.27)>
    <Clock(total=2.74, wait=1.87, extend-volume=0.58, refresh-volume=0.29)>

I tested also a shorter eventInterval (0.2 seconds). This reduces the
extend time by 20%, but doubles the cpu usage on the SPM mailbox.

With this change, the next improvement is eliminating the wait time. In
the worst case, we waited 1.97 seconds before sending an extend request,
which is 68% of the total extend time (3.31 seconds). This issue is
tracked in #85.

Fixes #102.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>