copy_data: only skip lock for div endpoint
We need to set lock_image if not both endpoint are DIV, and we need to
check img_id only if both are DIV, otherwise this operation will fail
with a KeyError as external endpoints do not have the img_id property.

This patch sets img_id to None for the external endpoint and checks
within the div endpoint if we need to lock the source image.

Bug-Url: https://bugzilla.redhat.com/2066285
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>