mailbox: Configurable event interval
Add "mailbox:events_interval" configuration, allowing using longer or
shorter events interval without changing the code. This is useful for
testing the best value in scale tests or to tuning in production.

To change the configuration, users can add a file on all hosts like:

    $ cat /etc/vdsm/vdsm.conf.d/99-local.conf
    [mailbox]
    events_interval = 0.25

And restart the vdsmd service.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>