tests: Refine mailbox rounddtrip timeouts
The roundtrip tests is fast enough now regardless of the number of
messages, and our CI is pretty good, so we don't need extreme timeouts.

Use time.monotonic() instead of time.time(), this is more reliable in
case system time changes during a run.

If a timeout expires, raise RuntimeError instead of assert. This is not
a test failure but an error.

Use EVENT_INTERVAL for sleeping, since when using smaller interval we
can detect cleared mailbox faster.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>