tests: Verify roundtrip stats
Verify that best, average and worst roundtrip times are expected. This
should protect from performance regressions in the mailbox.

Testing on github shows that the tests can be 2 times slower compared
with local run. We use larger timeouts to avoid random failures.

The docstring explains how to run the roundript tests and inspect the
results manually.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>